gl_basic_test: Run platform tests in separate process
In gl_basic_tests, run the each platform's testsuite in a forked process.
If any testsuite fails, don't run any more testsuites.

This isn't really needed for running glx and x11_egl. They can safely run
in the same process. But it will not be safe to run x11_egl and wayland
tests in the same process. Mesa's libEGL is able to set the EGL platform
only once per process.

Signed-off-by: Chad Versace <chad.versace@linux.intel.com>