include: Add new enums WAFFLE_CONTEXT_* and document
This patch mostly consists of documentation in waffle_config.h for the new
enums.

The new enums are not yet used. Some refactoring is needed before that is
possible.

The new context attributes are associated with waffle_config_choose rather
than waffle_context_create because, on Apple, the context profile
attributes must be given to CGLChoosePixelFormat.

Signed-off-by: Chad Versace <chad.versace@linux.intel.com>