core: If multiple errors are emitted, wisely choose a winner
Waffle is incapable of emitting a sequence of errors. So, if multiple
errors are emitted, Waffle has to choose a winner.

The first error emitted will likely be the most significant one in the
sequence, so let the first one win unless an internal error occurrs.
Internal errors always get priority.

The following unit tests in the wcore_error suite have been renamed and
fixed in tandem with this change:
    last_call_without_message_wins -> first_call_without_message_wins
    last_call_with_message_wins    -> first_call_with_message_wins

Signed-off-by: Chad Versace <chad.versace@linux.intel.com>