include: Kill waffle_finish()
I originally had high hopes for this function, but it failed to work as
expected, but by no fault of Waffle.

I had wanted to be able to initialize Waffle multiple times in one
process, each time with a different platform: GLX, Wayland, then X11/EGL.
However, Mesa's libEGL can only be initialized once per process.

Waffle shouldn't pretend to the user that it can do the impossible, so
just remove waffle_finish() and all documentation that referes to it and
re-initialization.

Signed-off-by: Chad Versace <chad.versace@linux.intel.com>