include,core: Let waffle_error_get_info() return a struct
Using waffle_error_get_info() was awkward. It had three return parameters.
To use it, the user had to declare three temporary variables into which
the info was returned.

This patch changes the function to return a struct that contains all the
info formerly returned in the output parameters.  This allows usage like:
    printf("ouch! %s\n", waffle_error_get_info()->message);

Some documentation in waffle_error.h is fixed up and relocated, as
appropriate with this change.

Several tests are fixed in tandem with this change.

Signed-off-by: Chad Versace <chad.versace@linux.intel.com>