glx: Fix glx_window, don't use glXCreateWindow
Fixes an X error produced by piglit_glx_window_set_no_input() in Piglit.

In GLX 1.2, one created a window and made it current like this:
    xcb_window_t window = xcb_generate_id(...);
    xcb_create_window(..., window, ...);
    glXMakeCurrent(..., window, ...);

In GLX 1.3, a new function was introduced: glXCreateWindow. The preferred
way became:
    xcb_window_t window = xcb_generate_id(...);
    xcb_create_window(..., window, ...);
    GLXWindow glx_window = glXCreateWindow(..., window, ...);
    glXMakeCurrent(..., glx_window, ...);

Waffle, up until this patch, used the GLX 1.3 method. However, `window`
and `glx_window` are *distinct* XID's. As discovered in Piglit, this
causes problems when other code needs to act on the window and uses the
wrong XID.

The solution: ensure that the window has a unique XID by using the GLX 1.2
method.

Fixes: piglit_glx_window_set_no_input()
Signed-off-by: Chad Versace <chad.versace@linux.intel.com>