wcore_error_unittest: Fix non-terminating loop
When compiled with `clang -O3`, the following code in
test_wcore_error_thread_local() is translated into a non-terminating loop:
    original source:
        while (num_threads_waiting < NUM_THREADS)
            ;;

    assembly:
        66d cmpl    $0x2,0x94(%rsp)
        674 noop
        675 jle     740
        ...
        740 jmp     740

The problem is that the compiler assumes that num_threads_waiting is
accessed only within the local scope and concludes that, if the "while"
condition fails on the loop's first iteration, then it fail for all
following iterations.  However, the assumption is false because others
threads asynchronously increment num_threads_waiting.

The solution is to declare num_threads_waiting as volatile. From the C99
spec:
    A volatile declaration may be used to describe [...] an object
    accessed by an asynchronously interrupting function. Actions on
    objects so declared shall not be ‘‘optimized out’’ by an
    implementation or reordered except as permitted by the rules for
    evaluating expressions.
This change forces the compiler to generate a memory access to read
num_threads_waiting on each iteration of the loop.  Below is the fixed
source and generated assembly.
    original source:
        while (num_threads_waiting < NUM_THREADS)
            ;;

    assembly:
        670 cmpl    $0x3,0x94(%rsp)
        674 noop
        678 jl      670

Signed-off-by: Chad Versace <chad.versace@linux.intel.com>