waffle: Use a static libwaffle for unittests
The unittest executable needs access to waffle's internal symbols. Yet, the
sources in libwaffle.so will soon be compiled with -fvisibility=hidden.
One solution is to additionally build a static libwaffle and let the
unittest executable link to that. Visibility options do not apply to static
libraries; that is, all symbols are public and hence available to the unit
tests.

Signed-off-by: Chad Versace <chad.versace@linux.intel.com>