pkg/gentoo: Add waffle-1.0.1.ebuild
Hey Chad,

Attached is an ebuild for waffle-1.0.1. The gles1/2 USE flags simply
ensure that Mesa is built with gles1/2 support. Gentoo should really
have a virtual for egl/gles1/gles2, but we don't yet.

We could add examples and test USE flags, but I'm not sure if we should.

Matt