tests/gl_basic: Disable GBM tests
The gl_basic testsuite can never pass if the gbm tests are run.

When a display manager, such as X or Weston, owns the drm device, it's
impossible to run the gbm tests as non-root. This makes `check-func`
always fail. It's possible to run the gbm tests as a normal user if no
display manager currently owns drm, but then the other tests fail.

Until we find a better solution for gbm initialization, it's best to leave
the gbm tests out and just manually tests it with `examples/gl_basic`.

Signed-off-by: Chad Versace <chad.versace@linux.intel.com>