core: Aggregate all object vtbl's into wcore_platform_vtbl
Functions have always been dispatched at the granularity of platform, not
of object. Having each object possess a vtbl is unnecessary and just
overcomplicates things. Virtual dispatch in C is messy business, and we
should keep it as simple as possible.

This patch moves the vtbl for each of wcore_display, wcore_config,
wcore_context, and wcore_window into the vtbl of wcore_platform.

Signed-off-by: Chad Versace <chad.versace@linux.intel.com>