egl: Rewrite egl module
In the beginning, there was only one EGL platform: xegl.  Later, wayland
and android and gbm were added. As each new platform was added, it
inherited a large portion of the previous platform's egl code by
copy-and-paste. Some egl code was shared via helper functions in
egl_no_native.c and egl_native_template.c, but still large portions of
code were duplicated.

The most egregious example of the code duplication took place in the
config and context objects. The files {xegl,wayland,droid,gbm}_config.[ch]
are are nearly identical boilerplate. Ditto for context.[ch].

Now that a new feature needs to be added to the EGL code, support for ES3,
it is an opportune time to remove the duplicate EGL code and refactor.

The extant EGL utilitiy functions are "thin"; they take real EGL types as
input. In order to kill the duplicate code in the config and context
classes, and to smoothly add ES3 support, the thin utility functions need
to be replaced with real classes.

This patch rewrites the egl module by introducing four classes:
    wegl_display
    wegl_config
    wegl_context
    wegl_window

This patch keeps the old egl code in-place so that platform modules can
migrate to the new egl code one at a time.

Signed-off-by: Chad Versace <chad.versace@linux.intel.com>