core: Store wcore_config_attrs in wcore_config
This removes the need for each platform's config object to transport
additional platform-specific information to the context and window.

In particular, wgbm_config currently transports the GBM_FORMAT
to wgbm_window. Once the gbm module is converted to use the egl module,
however, that transport will be impossible without this patch because
wegl_config contains no platform-specific data.

Signed-off-by: Chad Versace <chad.versace@linux.intel.com>