api,core: Allow wcore_platform get_native func ptrs to be null
Some platforms, such as cgl and android, don't yet support any of the
get_native functions. It's silly for each of the two platforms to define
an identical get_native function for each public object, a function that
just emits WAFFLE_ERROR_UNSUPPORTED_ON_PLATFORM.

If the function pointer is null, then just detect that at the api entry
point and emit the error there.

Signed-off-by: Chad Versace <chad.versace@linux.intel.com>