pkg/archlinux: Generate real PKGBUILD from template
This patch is a conceptual revert of 3245b12. It adds a new cmake rule
that generates a PKGBUILD from PKGBUILD.template.

Ken told me that he expected the build system to do this generation for
him, and he was surprised when he discovered that it didn't.

Requested-by: Kenneth Graunke <kenneth@whitecape.org>
Signed-off-by: Chad Versace <chad.versace@linux.intel.com>
(cherry picked from commit 0a16e5ef3ca36451e752449579c79dad20606f8b)