core: Rewrite wcore_config_attrs_parse()
Rewrite in order to interleave assignment of default values and the
parsing of the attrib list for different attribute groups. No behavioral
change is expected.

Before, wcore_config_attrs_parse() did these steps:
    - set default misc attributes
    - parse api
    - set default context version
    - parse and validate all attributes but api

Now wcore_config_attrs_parse() does these steps:
    - validate all key names
    - parse api
    - set default context version then parse context version
    - set default profile then parse profile
    - set default misc attributes and parse

The prompt for this patch is that the next commit alters the behavior of
the profile attribute in a way that requires the interleaving.

Signed-off-by: Chad Versace <chad.versace@linux.intel.com>