include: Fold nearly all headers into waffle.h
Fold all headers into waffle.h except
    - The native platform headers, because users need to be able to select
      which to include and exclude.
    - waffle_version.h, which is generate at configure time.

When Waffle's documentation consisted of Doxygen in the headers, it was
sensible to maintain a separate header for each class of functions. The
Doxygen was lengthy, thus having separate headers made the documentation
easier to navigate.

Now that the documentation resides in manpage xml, each individual header
is nearly empty. It's silly to have numerous headers that declare only
3 to 4 functions each.

Strictly speaking, removing header files breaks the API. But I'm aware of
no client that will miss the removed headers.

Signed-off-by: Chad Versace <chad.versace@linux.intel.com>