glx: Add glXCreateContextAttribs workaround for NVidia bug
With NVidia, calling glXCreateContextAttribsARB with MAJOR=1 and MINOR=0
returns an OpenGL 2.1 context. Calling it with MAJOR and MINOR unspecified
returns a context of the latest supported OpenGL version. However, according to
the GLX_ARB_create_context spec, the two calls should behave identically.
Calling with MAJOR and MINOR unspecified produces the expected behavior.

As a workaround, do not specify GLX_CONTEXT_MAJOR_VERSION_ARB and
GLX_CONTEXT_MINOR_VERSION_ARB in the call to glXCreateContextAttribsARB if the
user requested an OpenGL context of unspecified version or if the user
explicitly requested an OpenGL 1.0 context.

Reported-by: Paul Berry <stereotype441@gmail.com>
Reviewed-and-tested-by: Paul Berry <stereotype441@gmail.com>
Signed-off-by: Chad Versace <chad.versace@linux.intel.com>