core,man: Fix range and default value of WAFFLE_*_SIZE attrs
For each of the WAFFLE_*_SIZE attributes:
  - change the default value from WAFFLE_DONT_CARE(-1) to 0, and
  - change the valid range of values from [-1, +inf) to [0, +inf).

This change aligns waffle's behavior to that of GLX [1], EGL [2], and (if
I interpreted the CGL documentation correctly) CGL [3].

This fixes a bug where, if the user didn't specify the value of a size
attribute, then waffle passed WAFFLE_DONT_CARE(-1) to glXChooseFBConfig
and eglChooseConfig. Per the GLX, EGL specs, -1 is an illegal value for
size attributes.

[1] GLX 1.4 spec (2005.12.16), Table 3.4
[2] EGL 1.4 spec (2011.04.06), Table 3.4
[3] https://developer.apple.com/library/mac/documentation/graphicsimaging/reference/CGL_OpenGL/Reference/reference.html#//apple_ref/c/func/CGLChoosePixelFormat

Reported-by: Brian Paul <brian.e.paul@gmail.com>
Signed-off-by: Chad Versace <chad.versace@linux.intel.com>