glx: Fix gles regression
In glx_context_fill_attrib_list, recently a check was added to
work around a bug for OpenGL v1.0 contexts and NVidia. This
check introduced a bug where for OpenGL ES (all versions)
the attrib_list wasn't getting entries for
GLX_CONTEXT_MAJOR_VERSION_ARB or GLX_CONTEXT_MINOR_VERSION_ARB.
In piglit, this causes all OpenGL ES testcases to fail with
context failures.
This fix adjusts the comparison so that OpenGL ES will always
have the context version and not set it for a Open GL v1.0
context.

[chadv: Replace != OPENGL_ES1 with == OPENGL to improve readability.
 Double negatives are confusing.]

Signed-off-by: Tom Gall <tom.gall@linaro.org>
Signed-off-by: Chad Versace <chad.versace@linux.intel.com>