core: Rewrite code for thread local storage
The original code in wcore_tinfo.c for handling thread local info was
a mess. It was my first time writing thread local code in a non-classroom
setting.  This patch rewrites wcore_tinfo.c, reducing the lines of code
from ~130 to ~50.

The rewrite does the following:
  - Require that the compiler support thread local storage through the
    keyword '__thread'. This allows us to remove use of pthread_key_t.
  - Remove all mutexes. There is no need to protect thread-local storage
    with mutexes.

Tested with `make check-func` on GLX, XEGL, and Wayland.

Signed-off-by: Chad Versace <chad.versace@linux.intel.com>