cmake: Fix test for thread-local storage
Commit 6fcbe52 "cmake: Refactor compiler flags into new module" exposed a bug
in the test for thread-local storage, causing the test to always fail. If the
regex "warning|error" appears in the compiler log, then the test fails. This
causes false failures when the test invokes the compiler with
'-Werror=missing-prototypes'.

Fix the regex to more closely match gcc and clang error messages.

Reported-by: Paul Berry <stereotype441@gmail.com>
Tested-by: Paul Berry <stereotype441@gmail.com>
Signed-off-by: Chad Versace <chad.versace@linux.intel.com>