cgl: Simulate EGL in handling context versions
Pre-patch, when built on Mac OS < 10.6, the requested context version
was required to be exactly 1.0; and when built on Mac OS >= 10.6,
exactly 1.0 or 3.2.

Post-patch, emulates EGL_KHR_create_context, which allows the
implementation to return a context of the latest supported flavor that
is backwards-compatibile with the requested flavor.

Piglit tests that request a 3.1 context will require this change.

Signed-off-by: Chad Versace <chad.versace@linux.intel.com>