waffle: Mark waffle_window_resize() as experimental
That is, guard it with the feature macro WAFFLE_API_EXPERIMENTAL.

waffle_window_resize() is proving to dificult to implement because
Waffle lacks an event model.

It seems to work properly on Wayland. It works on CGL but the resize is
janky. It kinda works on composited X11 but interacts badly with GL
rendering; the client needs to stall after resize to provide the
compositor sufficient time to process the request before the client
begins to draw... ouch. I haven't tested it on Android.

So... until waffle_window_resize() actually works it needs to remain
experimental until we do one of:

    1. Make it synchronous on all platforms. (I'm having difficulty
       doing this on composited X11).

    2. Waffle exposes API for an event model, into which window resizes
       are integrated. (This is likely the wrong direction).

    3. We kill waffle_window_resize().

CC: Juha-Pekka Heikkila <juha-pekka.heikkila@linux.intel.com>
Signed-off-by: Chad Versace <chad.versace@linux.intel.com>