cmake: Move targets check, check-func to top dir
By moving 'check' to the top dir, each subdirectory can now add
dependencies to 'check'. Ditto for 'check-func'.

This change will allow us to build each unittest file as a separate
executable, as well as maintain the CMake rules for the unittest in in
the same directory as its sources.

Signed-off-by: Chad Versace <chad.versace@linux.intel.com>