cmake: Fix dependency on cmocka
This fixes an error in the Makefile-dependency of unittests, introduced
in commit ad40427a864a7436953b319fbfdb2c04ccb7bdab.

From the perspective of Waffle's CMake, libcmocka.so is an *imported*
library. In other words, it's a virtual target, represented by
a '.PHONY' target in Unix Makefiles.

To guarantee that libcmocka.so gets built, CMake must declare that the
.PHONY 'cmocka' target depends on another .PHONY target, which does the
actual building. Making 'cmocka' depend on a non-PHONY target doesn't
work.

Reported-by: Shazia Rahman <srahman@nvidia.com>
Reported-by: Paul Berry <stereotype441@gmail.com>
Tested-by: Paul Berry <stereotype441@gmail.com>
Signed-off-by: Chad Versace <chad.versace@linux.intel.com>