wayland: add a handler for global remove event
The Wayland event dispatch infrastructure does not check if the event
handler member is NULL before dereferencing the appropriate member in
the listener.

This change supplies an empty function as the event handler for the
global remove event on the wl_registry listener rather than the
previously supplied NULL value.

Signed-off-by: Rob Bradford <rob@linux.intel.com>
Signed-off-by: Chad Versace <chad.versace@linux.intel.com>