cgl: Make waffle_get_proc_address() work
There is no CGLGetProcAddress, so waffle_get_proc_address() always
emitted an error.  However, this patch changes it to follow the
principle of least surprise. The only supported API on CGL is OpenGL, so
assume the user called waffle_get_proc_address() to obtain an OpenGL
function pointer and get it with dlsym().

Signed-off-by: Chad Versace <chad.versace@linux.intel.com>