utils/wflinfo: allow requesting a compat profile without a version
Similar to the behavior of --profile=core, we will now
scan a list of 'known' profile supporting GL versions
when the user requests a compat profile, but does not
specify a version.

Tested on the NVIDIA 310.44 driver. It exposes GL_ARB_compatibility
as expected when --profile=compat is used, and does not when
--profile=core is used.

Signed-off-by: Jordan Justen <jljusten@gmail.com>
Signed-off-by: Chad Versace <chad.versace@linux.intel.com>