waffle: Set LINK_INTERFACE_LIBRARIES to empty
Debian's packaging tools show a warning with wflinfo, because it
links to several libraries that it doesn't directly use. (libEGL,
libgbm, libGL, libudev, libwayland-client, ...)

On http://www.cmake.org/Wiki/CMake_FAQ
under "Why are libraries linked to my shared library included when
something links to it?"

"This question arises when one has a library B which links to some
 library A. When a third target, say C, links to B, CMake will
 automatically include C to A also."

The same FAQ item indicates that setting LINK_INTERFACE_LIBRARIES
to "" will disable this behavior.

Note: lddtree in pax-utils was also useful in showing a link tree
for wflinfo.

Signed-off-by: Jordan Justen <jordan.l.justen@intel.com>
Cc: Carl Worth <cworth@cworth.org>
Signed-off-by: Chad Versace <chad.versace@linux.intel.com>