waffle: remove restrict keyword from public API
We should have never defined the restrict keyword in
the waffle.h header. The word is used as a hint to the
compiler to provide optimisations re: pointer aliasing.

Keep the symbol around, so that the compiler can still
optimise the library.

v2: Move the symbol definition to its user and update the
man pages.

Resolves #11: https://github.com/waffle-gl/waffle/issues/11
Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>
Reviewed-by: Chad Versace <chad.versace@linux.intel.com>