wgl: wire-up wgl_window and wgl_config hooks
WGL mandates that a window (device_context actually) must be
available before we choose the the config and create a context.
Create the window at wgl_config_choose() and hide it until
wgl_window_create() is executed.

Note that this will mandate 'only one window is allowed per config'
as such one should _never_ try to create two windows for a single
config with WGL.

To prevent issues, simply return NULL if we're trying to
create the second window.

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>
Reviewed-by: Jose Fonseca <jfonseca@vmware.com>
Acked-by: Chad Versace <chad.versace@linux.intel.com>