linux: plug a memory leak
In linux_platform_destroy we close all the handles for the different
libgl* providers, yet we do not free the memory for the base struct
linux_platform, causing a memory leak.

In reality the memory will never be freed ('still reachable' by valgrind)
as we do not have a API to call platform->destroy() and thus cleanup
this and other memory hunks.

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>
Reviewed-by: Brian Paul <brianp@vmware.com>
Reviewed-by: Jose Fonseca <jfonseca@vmware.com>
Reviewed-by: Chad Versace <chad.versace@linux.intel.com>