glx: Don't use ARB_create_context with pre 3.2 contexts
This way if the user requests GL pre 3.2 context which lacks the
flags/extra bits which require ARB_create_context one can safely fall
back to the normal/legacy entry point.

This resolves piglits on non 3.2 capable drivers such as classic swrast,
nouveau_vieux and alike.

v2: The version should be 3.2 and not 3.0 as originally.

Cc: Ilia Mirkin <imirkin@alum.mit.edu>
Reviewed-by: Jose Fonseca <jfonseca@vmware.com> (v1)
Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>
Signed-off-by: Chad Versace <chad.versace@intel.com>