gbm: implement window_resize
Reuse the existing init/teardown functions to create a new window and
nuke the old one on success.

Note: as we need the original config, we keep a reference to it in
struct wcore_config. Ideally there will be a generic way to query it.

v2: attribute the wegl_window -> wegl_surface rename

Signed-off-by: Emil Velikov <emil.velikov@collabora.com>