Revert "wegl: add EGL image create/destroy"
This reverts commit 1a8e8ef3b8072fa1135dbe2a04dc6bb9f8e3a12f.

The symbols were added as they are required by the null/surfaceless
platform. Since that one hasn't landed yet, just remove the patch.

Furthermore, using dlsym is non-compatible (according to the spec) since
the symbols are part of an extension. eglGetProcAddress should be used.

Feel free to bring it back, when needed.

Cc: Frank Henigman <fjhenigman@google.com>
Reviewed-by: Chad Versace <chadversary@chromium.org>