egl: Use eglGetPlatformDisplayEXT as a fallback
If the EGL platform does not support eglGetPlatformDisplay, but does
support eglGetPlatformDisplayEXT, then use it as a fallback. Of course,
if neither is supported, we fall back all the way to eglGetDisplay.

This is useful for exercising the codepaths on Mesa, which does support
the extensions for eglGetPlatformDisplayEXT but not yet
eglGetPlatformDisplay.

Tested against Mesa master@8c78fdb with `ninja check-func` on Linux for
x11_egl, wayland, and gbm.

Reported-by: Emil Velikov <emil.l.velikov@gmail.com>
Reviewed-by: Emil Velikov <emil.l.velikov@gmail.com>