Initial meson build system
This provides all of the capacities of the cmake build system except for
the find files (those are pretty complicated and specific to cmake), and
support for nacl. I don't have any way to test nacl (although there is
some code in the meson build to support it), and the nacl code we have
seems to be fairly out of date. I'm ready to merge this now and deal
with nacl later.

Done:
 - linux
 - tests
 - examples
 - docs
 - mac
 - install
 - use with piglit
 - windows

Todo:
 - nacl?

Reviewed-by: Eric Engestrom <eric.engestrom@intel.com>