debian: Remove debian from upstream waffle tree
The debian project actually prefers that external upstream projects
not include debian packaging. The reason is that the debian project
prefers to have ownership over the packaging in case a debian
developer needs to update the packaging, but doesn't have access to
the upstream repo.

This is documented on the wiki:

https://wiki.debian.org/DebianMentorsFaq#What.27s_wrong_with_upstream_shipping_a_debian.2F_directory.3F

Signed-off-by: Jordan Justen <jordan.l.justen@intel.com>
Reviewed-by: Dylan Baker <dylan@pnwbakers.com>