wflinfo: use glGetStringi() only when available
In some corner cases, we may end up attempting to get the extension list
via the new API, even though we're yet to get function pointer.

Fetching symbols is fun, as can be seen in the beefy comment above the
glGetStringi retrieval.

With this in place, we no longer crash with the following:

  export LIBGL_ALWAYS_SOFTWARE=1
  wflinfo --platform glx --api gl --profile compat --version 3.1

Reported-by: Ben Crocker <bcrocker@redhat.com>
Signed-off-by: Emil Velikov <emil.velikov@collabora.com>