meson: use override_dependency and override_find_program
This allows automatic overrides of the waffle-1 dependency and wflifo
program, when used as a subproject/dependency fallback without having to
specify them in the consumer, as is current meson best practice.

For projects using meson > 0.54, they can then just write:

```meson
dependency('waffle-1')
```
and if they have the waffle-1 wrap installed then meson will
automatically configure the project, and use the built libwaffle-1,
without having to specify fallback parameters in the calling
meson.build. This is quite similar to the find_program_override, which
works for executables instead of dependencies.